Changed type of variable n in camtipCorr.hpp from size_t to uint64_t. Added a "pragma once" statement to the top of cc-functions.hpp

+++ apps/camtipCorr/camtipCorr.hpp
@@ -160,7 +160,7 @@ class camtipCorr : public MagAOXApp<false>,
       //m_means[2] holds mean Strehl Ratios 'SR' 
       std::array<std::array<realT, 2>, 3> m_means;
 
-      size_t n;
+      uint64_t n;
 };
